+++ toolkit/components/telemetry/build_scripts/mozparsers/shared_telemetry_utils.py
@@ -26,15 +26,13 @@ KNOWN_PROCESS_FLAGS = {
     "all_childs": "AllChildren",  # Supporting files from before bug 1363725
 }
 
-GECKOVIEW_STREAMING_PRODUCT = "geckoview_streaming"
-
 SUPPORTED_PRODUCTS = {
     "firefox": "Firefox",
     "fennec": "Fennec",
-    GECKOVIEW_STREAMING_PRODUCT: "GeckoviewStreaming",
     "thunderbird": "Thunderbird",
     # Historical, deprecated values:
     # 'geckoview': 'Geckoview',
+    # "geckoview_streaming": "GeckoviewStreaming",
 }
 
 SUPPORTED_OPERATING_SYSTEMS = [
@@ -112,10 +110,6 @@ def process_name_to_enum(name):
 
 def is_valid_product(name):
     return name in SUPPORTED_PRODUCTS
-
-
-def is_geckoview_streaming_product(name):
-    return name == GECKOVIEW_STREAMING_PRODUCT
 
 
 def is_valid_os(name):