+++ toolkit/components/telemetry/core/TelemetryCommon.h
@@ -42,10 +42,8 @@ static_assert(static_cast<uint16_t>(RecordedProcessType::Main) == 1,
 enum class SupportedProduct : uint8_t {
   Firefox = (1 << 0),
   Fennec = (1 << 1),
-  // Note that `1 << 2` (former GeckoView) is missing in the representation
-  // but isn't necessary to be maintained, but we see no point in filling it
-  // at this time.
-  GeckoviewStreaming = (1 << 3),
+  // Note that `1 << 2` and `1 << 3` (former GeckoView, GeckoviewStreaming) are
+  // missing in the representation. We see no point in filling it at this time.
   Thunderbird = (1 << 4),
 };
 MOZ_MAKE_ENUM_CLASS_BITWISE_OPERATORS(SupportedProduct);