+++ toolkit/components/telemetry/core/TelemetryHistogram.h
@@ -12,11 +12,6 @@
 #include "nsXULAppAPI.h"
 #include "TelemetryCommon.h"
 
-namespace mozilla {
-// This is only used for the GeckoView persistence.
-class JSONWriter;
-}  // namespace mozilla
-
 // This module is internal to Telemetry.  It encapsulates Telemetry's
 // histogram accumulation and storage logic.  It should only be used by
 // Telemetry.cpp.  These functions should not be used anywhere else.
@@ -109,15 +104,6 @@ nsresult GetKeyedHistogramSnapshots(JSContext* aCx,
                                     bool aFilterTest = false);
 
 size_t GetHistogramSizesOfIncludingThis(mozilla::MallocSizeOf aMallocSizeOf);
-
-// These functions are only meant to be used for GeckoView persistence.
-// They are responsible for updating in-memory probes with the data persisted
-// on the disk and vice-versa.
-nsresult SerializeHistograms(mozilla::JSONWriter& aWriter);
-nsresult SerializeKeyedHistograms(mozilla::JSONWriter& aWriter);
-nsresult DeserializeHistograms(JSContext* aCx, JS::Handle<JS::Value> aData);
-nsresult DeserializeKeyedHistograms(JSContext* aCx,
-                                    JS::Handle<JS::Value> aData);
 
 }  // namespace TelemetryHistogram