+++ toolkit/components/telemetry/tests/python/test_parse_events.py
@@ -145,22 +145,6 @@ expiry_version: never
             ),
         )
 
-    def test_geckoview_streaming_product(self):
-        SAMPLE_EVENT = """
-methods: ["method1", "method2"]
-objects: ["object1", "object2"]
-bug_numbers: [12345]
-notification_emails: ["test01@mozilla.com", "test02@mozilla.com"]
-record_in_processes: ["content"]
-description: This is a test entry for Telemetry.
-products: ["geckoview_streaming"]
-expiry_version: never
-"""
-        event = load_event(SAMPLE_EVENT)
-        parse_events.EventData("CATEGORY", "test_event", event, strict_type_checks=True)
-
-        self.assertRaises(SystemExit, ParserError.exit_func)
-
 
 if __name__ == "__main__":
     mozunit.main()